+++ src/dragdrop/gridstack-dd-native.ts
@@ -6,11 +6,12 @@
  * gridstack.js may be freely distributed under the MIT license.
 */
 import { DDManager } from './dd-manager';
-import { DDElement } from './dd-element';
+import { DDElement, DDElementHost } from './dd-element';
 
 import { GridStackElement } from '../gridstack';
 import { GridStackDD, DDOpts, DDKey, DDDropOpt, DDCallback, DDValue } from '../gridstack-dd';
 import { GridItemHTMLElement, DDDragInOpt } from '../types';
+import { Utils } from '../utils';
 
 /**
  * HTML 5 Native DragDrop based drag'n'drop plugin.
@@ -18,119 +19,117 @@ import { GridItemHTMLElement, DDDragInOpt } from '../types';
 export class GridStackDDNative extends GridStackDD {
 
   public resizable(el: GridItemHTMLElement, opts: DDOpts, key?: DDKey, value?: DDValue): GridStackDDNative {
-    let dEl = this.getGridStackDDElement(el);
-    if (opts === 'disable' || opts === 'enable') {
-      dEl.ddResizable[opts]();
-    } else if (opts === 'destroy') {
-      if (dEl.ddResizable) {
-        dEl.cleanResizable();
-      }
-    } else if (opts === 'option') {
-      dEl.setupResizable({ [key]: value });
-    } else {
-      const grid = el.gridstackNode.grid;
-      let handles = dEl.el.getAttribute('gs-resize-handles') ? dEl.el.getAttribute('gs-resize-handles') : grid.opts.resizable.handles;
-      dEl.setupResizable({
-        ...grid.opts.resizable,
-        ...{ handles: handles },
-        ...{
-          start: opts.start,
-          stop: opts.stop,
-          resize: opts.resize
+    this.getDDElements(el).forEach(dEl => {
+      if (opts === 'disable' || opts === 'enable') {
+        dEl.ddResizable[opts]();
+      } else if (opts === 'destroy') {
+        if (dEl.ddResizable) {
+          dEl.cleanResizable();
         }
-      });
-    }
-    return this;
-  }
-
-  public draggable(el: GridItemHTMLElement, opts: DDOpts, key?: DDKey, value?: DDValue): GridStackDDNative {
-    const dEl = this.getGridStackDDElement(el);
-    if (opts === 'disable' || opts === 'enable') {
-      dEl.ddDraggable && dEl.ddDraggable[opts]();
-    } else if (opts === 'destroy') {
-      if (dEl.ddDraggable) { // error to call destroy if not there
-        dEl.cleanDraggable();
+      } else if (opts === 'option') {
+        dEl.setupResizable({ [key]: value });
+      } else {
+        const grid = dEl.el.gridstackNode.grid;
+        let handles = dEl.el.getAttribute('gs-resize-handles') ? dEl.el.getAttribute('gs-resize-handles') : grid.opts.resizable.handles;
+        dEl.setupResizable({
+          ...grid.opts.resizable,
+          ...{ handles: handles },
+          ...{
+            start: opts.start,
+            stop: opts.stop,
+            resize: opts.resize
+          }
+        });
       }
-    } else if (opts === 'option') {
-      dEl.setupDraggable({ [key]: value });
-    } else {
-      const grid = el.gridstackNode.grid;
-      dEl.setupDraggable({
-        ...grid.opts.draggable,
-        ...{
-          containment: (grid.opts._isNested && !grid.opts.dragOut)
-            ? grid.el.parentElement
-            : (grid.opts.draggable.containment || null),
-          start: opts.start,
-          stop: opts.stop,
-          drag: opts.drag
-        }
-      });
-    }
-    return this;
-  }
-
-  public dragIn(el: GridStackElement, opts: DDDragInOpt): GridStackDDNative {
-    let dEl = this.getGridStackDDElement(el);
-    dEl.setupDraggable(opts);
-    return this;
-  }
-
-  public droppable(el: GridItemHTMLElement, opts: DDOpts | DDDropOpt, key?: DDKey, value?: DDValue): GridStackDDNative {
-    let dEl = this.getGridStackDDElement(el);
-    if (typeof opts.accept === 'function' && !opts._accept) {
-      opts._accept = opts.accept;
-      opts.accept = (el) => opts._accept(el);
-    }
-    if (opts === 'disable' || opts === 'enable') {
-      dEl.ddDroppable && dEl.ddDroppable[opts]();
-    } else if (opts === 'destroy') {
-      if (dEl.ddDroppable) { // error to call destroy if not there
-        dEl.cleanDroppable();
-      }
-    } else if (opts === 'option') {
-      dEl.setupDroppable({ [key]: value });
-    } else {
-      dEl.setupDroppable(opts);
-    }
-    return this;
-  }
-
-  public isDroppable(el: GridItemHTMLElement): boolean {
-    const dEl = this.getGridStackDDElement(el);
-    return !!(dEl.ddDroppable);
-  }
-
-  public isDraggable(el: GridStackElement): boolean {
-    const dEl = this.getGridStackDDElement(el);
-    return !!(dEl.ddDraggable);
-  }
-
-  public on(el: GridItemHTMLElement, name: string, callback: DDCallback): GridStackDDNative {
-    let dEl = this.getGridStackDDElement(el);
-    dEl.on(name, (event: Event) => {
-      callback(
-        event,
-        DDManager.dragElement ? DDManager.dragElement.el : event.target as GridItemHTMLElement,
-        DDManager.dragElement ? DDManager.dragElement.helper : null)
     });
     return this;
   }
 
-  public off(el: GridItemHTMLElement, name: string): GridStackDD {
-    let dEl = this.getGridStackDDElement(el);
-    dEl.off(name);
+  public draggable(el: GridItemHTMLElement, opts: DDOpts, key?: DDKey, value?: DDValue): GridStackDDNative {
+    this.getDDElements(el).forEach(dEl => {
+      if (opts === 'disable' || opts === 'enable') {
+        dEl.ddDraggable && dEl.ddDraggable[opts]();
+      } else if (opts === 'destroy') {
+        if (dEl.ddDraggable) { // error to call destroy if not there
+          dEl.cleanDraggable();
+        }
+      } else if (opts === 'option') {
+        dEl.setupDraggable({ [key]: value });
+      } else {
+        const grid = dEl.el.gridstackNode.grid;
+        dEl.setupDraggable({
+          ...grid.opts.draggable,
+          ...{
+            containment: (grid.opts._isNested && !grid.opts.dragOut)
+              ? grid.el.parentElement
+              : (grid.opts.draggable.containment || null),
+            start: opts.start,
+            stop: opts.stop,
+            drag: opts.drag
+          }
+        });
+      }
+    });
     return this;
   }
 
-  private getGridStackDDElement(el: GridStackElement): DDElement {
-    let dEl;
-    if (typeof el === 'string') {
-      dEl =  document.querySelector(el as string);
-    } else {
-      dEl = el;
+  public dragIn(el: GridStackElement, opts: DDDragInOpt): GridStackDDNative {
+    this.getDDElements(el).forEach(dEl => dEl.setupDraggable(opts));
+    return this;
+  }
+
+  public droppable(el: GridItemHTMLElement, opts: DDOpts | DDDropOpt, key?: DDKey, value?: DDValue): GridStackDDNative {
+    if (typeof opts.accept === 'function' && !opts._accept) {
+      opts._accept = opts.accept;
+      opts.accept = (el) => opts._accept(el);
     }
-    return dEl.ddElement ? dEl.ddElement: DDElement.init(dEl);
+    this.getDDElements(el).forEach(dEl => {
+      if (opts === 'disable' || opts === 'enable') {
+        dEl.ddDroppable && dEl.ddDroppable[opts]();
+      } else if (opts === 'destroy') {
+        if (dEl.ddDroppable) { // error to call destroy if not there
+          dEl.cleanDroppable();
+        }
+      } else if (opts === 'option') {
+        dEl.setupDroppable({ [key]: value });
+      } else {
+        dEl.setupDroppable(opts);
+      }
+    });
+    return this;
+  }
+
+  /** true if at least one of them is droppable */
+  public isDroppable(el: GridItemHTMLElement): boolean {
+    return this.getDDElements(el).some(dEl => !!(dEl.ddDroppable));
+  }
+
+  /** true if at least one of them is draggable */
+  public isDraggable(el: GridStackElement): boolean {
+    return this.getDDElements(el).some(dEl => !!(dEl.ddDraggable));
+  }
+
+  public on(el: GridItemHTMLElement, name: string, callback: DDCallback): GridStackDDNative {
+    this.getDDElements(el).forEach(dEl =>
+      dEl.on(name, (event: Event) => {
+        callback(
+          event,
+          DDManager.dragElement ? DDManager.dragElement.el : event.target as GridItemHTMLElement,
+          DDManager.dragElement ? DDManager.dragElement.helper : null)
+      })
+    );
+    return this;
+  }
+
+  public off(el: GridItemHTMLElement, name: string): GridStackDD {
+    this.getDDElements(el).forEach(dEl => dEl.off(name));
+    return this;
+  }
+
+  private getDDElements(els: GridStackElement): DDElement[] {
+    let list = Utils.getElements(els) as DDElementHost[];
+    if (!list.length) { return []; }
+    return list.map(e => e.ddElement || DDElement.init(e));
   }
 }
